+++ src/main/java/bot/models/Operator.java
@@ -1,0 +1,5 @@
+package bot.models;
+
+public enum Operator {
+    ADD, SUBTRACT, MULTIPLY, DIVIDE, MODULO, AND, OR, XOR;
+}