+++ src/main/java/bot/models/Question.java
@@ -1,0 +1,7 @@
+package bot.models;
+
+public interface Question {
+
+    String getQuestion();
+    boolean checkAnswer(String answer);
+}